# swagger.yaml
openapi: 3.0.0
info:
  title: Mi API - Dip Mod 4 Juan Pablo Guzman
  version: 1.0.0
  description: Documentación de la API
  contact:
    name: Juan Guzman
    email: juan@email.com
    url: https://github.com/tu-usuario
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Desarrollo
  - url: https://tu-app.onrender.com
    description: Producción
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Juan Guzman
        email:
          type: string
          format: email
          example: juan@email.com
        active:
          type: boolean
          example: true
    UserInput:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: Juan Guzman
        email:
          type: string
          format: email
          example: juan@email.com
        password:
          type: string
          format: password
          example: '123456'
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Mi tarea
        description:
          type: string
          example: Descripción de la tarea
        done:
          type: boolean
          example: false
    TaskInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: Mi tarea
        description:
          type: string
          example: Descripción de la tarea
paths:
  /api/login:
    post:
      summary: Iniciar sesión
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: usuario@email.com
                password:
                  type: string
                  format: password
                  example: '123456'
      responses:
        '200':
          description: Login exitoso, retorna el token JWT
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Datos inválidos
        '401':
          description: Credenciales incorrectas
  /api/tasks:
    get:
      summary: Obtener todas las tareas
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de tareas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: No autorizado
    post:
      summary: Crear una nueva tarea
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '201':
          description: Tarea creada exitosamente
        '400':
          description: Datos inválidos
        '401':
          description: No autorizado
  /api/tasks/{id}:
    get:
      summary: Obtener una tarea por ID
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID de la tarea
      responses:
        '200':
          description: Tarea encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: No autorizado
        '404':
          description: Tarea no encontrada
    put:
      summary: Actualizar una tarea completa
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '200':
          description: Tarea actualizada
        '401':
          description: No autorizado
        '404':
          description: Tarea no encontrada
    delete:
      summary: Eliminar una tarea
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tarea eliminada
        '401':
          description: No autorizado
        '404':
          description: Tarea no encontrada
    patch:
      summary: Marcar tarea como completada
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tarea marcada como completada
        '401':
          description: No autorizado
        '404':
          description: Tarea no encontrada
  /api/users:
    get:
      summary: Obtener todos los usuarios
      tags:
        - Users
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Crear un nuevo usuario
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: Usuario creado exitosamente
        '400':
          description: Datos inválidos
  /api/users/list/pagination:
    get:
      summary: Listar usuarios con paginación
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            example: 1
          description: Número de página
        - in: query
          name: limit
          schema:
            type: integer
            example: 10
          description: Cantidad de resultados por página
      responses:
        '200':
          description: Lista paginada de usuarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 50
                  page:
                    type: integer
                    example: 1
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          description: No autorizado
  /api/users/{id}:
    get:
      summary: Obtener un usuario por ID
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del usuario
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: No autorizado
        '404':
          description: Usuario no encontrado
    put:
      summary: Actualizar un usuario completo
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: Usuario actualizado
        '400':
          description: Datos inválidos
        '401':
          description: No autorizado
        '404':
          description: Usuario no encontrado
    patch:
      summary: Activar o desactivar un usuario
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Estado del usuario actualizado
        '401':
          description: No autorizado
        '404':
          description: Usuario no encontrado
    delete:
      summary: Eliminar un usuario
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario eliminado
        '401':
          description: No autorizado
        '404':
          description: Usuario no encontrado
  /api/users/{id}/tasks:
    get:
      summary: Obtener las tareas de un usuario
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del usuario
      responses:
        '200':
          description: Lista de tareas del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: No autorizado
        '404':
          description: Usuario no encontrado
tags: []